pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
flask = "^3.0.0"
flask-session = "^0.5.0"
paho-mqtt = "^1.6.1"
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.1"
waitress = "^2.1.2"
werkzeug = "^3.0.0"
flask-wtf = "^1.2.1"
pycryptodome = "^3.19.0"
pyotp = "^2.9.0"
qrcode = "^7.4.2"
websockets = "^11.0.3"
starlette = "^0.31.1"
uvicorn = "^0.23.2"
sse-starlette = "^1.6.5"
pillow = "^10.1.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"